+++ opentelemetry-sdk/src/lib.rs
@@ -86,8 +86,6 @@
 //!
 //! For `logs` the following feature flags are available:
 //!
-//! * `spec_unstable_logs_enabled`: control the log level
-//!
 //! Support for recording and exporting telemetry asynchronously and perform
 //! metrics aggregation can be added via the following flags:
 //!