+++ opentelemetry-appender-tracing/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 ## vNext
 
+## 0.31.1
+
+Released 2025-Oct-1
+
+- Bump `tracing-opentelemetry` to 0.32
+
+## 0.31.0
+
+Released 2025-Sep-25
+
+- Updated `opentelemetry` dependency to version 0.31.0.
+
 ## 0.30.1
 
 Released 2025-June-05